# Optimize migration_cache.Read