@vanja-p 2025-05-15
> this is strictly better so let's get it in -- but we could look at using bytebufferpool here too

I thought about it, but it's way more complicated:
- bytebufferpool.VariableSize requires a max size, but we don't know what the max size will be, so we have to handle reads bigger than that by allocating a buffer ourselves.
- VariableSize creates log2(max size) sync.Pools, but for small reads we just need one buffer
- Every time we get a read bigger than the previously largest one, we would have to return the buffer to the pool and get another buffer.